pyproject.toml
[project]
name = "W2VNumpy"
version = "0.1.0"
description = "NumPy implementations of SGNS and GloVe on WikiText-2 with benchmarking."
authors = [{ name = "Alex Raudvee" }]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy",
    "hydra-core",
    "tqdm",
    "matplotlib",
    "scipy",
    "pytest",
    "requests",
    "datasets>=4.6.1",
    "scikit-learn>=1.7.2",
]

[project.optional-dependencies]
dev = ["ipykernel", "notebook"]

[tool.uv]
# uv-specific settings if you want any later